Added Yocto CMake Project Creation Wizard.  Removed old on based upon C/C++ templates.  Refactoring.
