+++ toolchains/sdk/org.yocto.crops.sdk.ui/src/org/yocto/crops/internal/sdk/ui/Activator.java
@@ -16,7 +16,7 @@ import org.osgi.framework.BundleContext;
 public class Activator extends AbstractUIPlugin {
 
 	// The plug-in ID
-	public static final String PLUGIN_ID = "org.yocto.crops.internal.sdk.ui"; //$NON-NLS-1$
+	public static final String PLUGIN_ID = "org.yocto.crops.sdk.ui"; //$NON-NLS-1$
 
 	// The shared instance
 	private static Activator plugin;
@@ -54,4 +54,5 @@ public class Activator extends AbstractUIPlugin {
 		return plugin;
 	}
 
+	
 }
